🔨 move Choropleth component to its own file
